show number of bookings and made admin mobile friendly

+++ seattlerealestatemedia/src/app/components/booking-list/booking-list.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import { customerRequest } from 'src/app/customerRequest';
-import { DatabaseService } from 'src/app/database.service';
+import { customerRequest } from 'src/app/data/models/customerRequest';
+import { DatabaseService } from 'src/app/data/services/database.service';
 
 @Component({
   selector: 'app-booking-list',
@@ -8,18 +8,39 @@ import { DatabaseService } from 'src/app/database.service';
   styleUrls: ['./booking-list.component.scss'],
 })
 export class BookingListComponent {
-  listOfBootings: customerRequest[] = [];
+  listOfBookings: customerRequest[] = [];
+  count: number = 0;
   constructor(private db: DatabaseService) {}
   ngOnInit() {
-    this.db.getBookings().subscribe((cr) => {
-      this.listOfBootings = cr as customerRequest[];
-    });
-    this.listOfBootings.forEach((element) => {
-      console.log(element);
-    });
+
+      this.db.getBookingsWithMetaData().subscribe({
+        next: (res) => {
+          //console.log('Next getting bookings ');
+        this.listOfBookings = res.map((book: any) =>{
+          const data = book.payload.doc.data();
+          data.id = book.payload.doc.id;
+          return data;
+
+        })
+        this.count = this.listOfBookings.length;
+      },
+      error:(err)  => {
+        console.log('Error getting bookings ' + err);
+        return err;
+        //alert('Error getting bookings ' + err)
+      },
+      complete:() =>{
+        //console.log('Complete getting bookings ');
+      }
+
+    })
+    this.count = this.listOfBookings.length;
   }
 
+
+
   onDelete(booking: customerRequest){
+    if(window.confirm("Are you sure you want to delete this booking " + booking.id +"?"))
     this.db.deleteBooking(booking);
   }
 }
